requirements.txt
alembic==1.10.2
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click== 8.1.3
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.3.1
fastapi==0.94.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
pip==23.0.1
pluggy==1.0.0
psycopg2-binary==2.9.5
pydantic==1.10.7
pytest==7.2.2
pytest-asyncio==0.20.3
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
setuptools==57.4.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
tomli==2.0.1
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
wheel==0.37.0
zipp==3.15.0
